+++ internal/server/directory.go
@@ -247,6 +247,8 @@ func rpcError(err error) error {
 		return connect.NewError(connect.CodeNotFound, err)
 	case errors.Is(err, hub.ErrDeclared):
 		return connect.NewError(connect.CodeFailedPrecondition, err)
+	case errors.Is(err, hub.ErrUnknownDomain):
+		return connect.NewError(connect.CodeInvalidArgument, err)
 	default:
 		return connect.NewError(connect.CodeInternal, err)
 	}